requirements.txt
python-dotenv==1.0.1
google-genai==0.3.0
pillow==10.2.0
streamlit==1.42.0
gTTS==2.5.4
fpdf2==2.8.4
